requirements/base.txt
whitenoise==6.6.0
Pillow==10.1.0
coverage==7.3.2
phonenumberslite==8.13.6
requests==2.31.0
six==1.16.0

#Django
django==4.2.7
django-environ==0.11.2 
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-filter==23.3
django-widget-tweaks==1.5.0
# Django Rest Framework
djangorestframework==3.14.0
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
django-extensions==3.2.3
